Make sure that new_index_value is captured before full_refresh (#70)
The new_index_value is captured lazily when the property is first
accessed. Unfortunately, for the full_refresh case this currently
doesn't happen until after the extraction completes which means
that the resulting index may potentially get ahead of the data,
which is bad. This change makes sure that full_refresh uses the
new_index_value within the query WHERE clause.

Also adds validation that a rebuild always also has a full_refresh.
There's logic in the table handling that makes this assumption
but wasn't explicitly validating it.